Update palindromedetector.py
Changed the test input type to a dictionary with the word/phrase as the key and the T/F of if it is actually a palindrome as the value for ease of testing. Also updated the output string formatting, added a standard "divider", and added a colon and a space to the end of the user input prompt.